% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Actions_Carrier_Drug}
\alias{Actions_Carrier_Drug}
\title{Drug Carriers related Actions}
\format{
a tibble with 15 observations and 2 variables:
\describe{
  \item{text}{describe related action}
  \item{parent_key}{carrier key}
}
}
\source{
\url{https://docs.drugbankplus.com/xml/}
}
\usage{
Actions_Carrier_Drug
}
\description{
A collection of actions related to drugs carriers
}
\keyword{datasets}
