% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Actions_Enzyme_Drug}
\alias{Actions_Enzyme_Drug}
\title{Drug Enzymes related Actions}
\format{
a tibble with 77 observations and 2 variables:
\describe{
  \item{text}{describe related action}
  \item{parent_key}{Enzyme key}
}
}
\source{
\url{https://docs.drugbankplus.com/xml/}
}
\usage{
Actions_Enzyme_Drug
}
\description{
A collection of actions related to drugs Enzymes
}
\keyword{datasets}
