% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Actions_Target_Drug}
\alias{Actions_Target_Drug}
\title{Drug Targets related Actions}
\format{
a tibble with 88 observations and 2 variables:
\describe{
  \item{text}{describe related action}
  \item{parent_id}{Target id}
}
}
\source{
\url{https://docs.drugbankplus.com/xml/}
}
\usage{
Actions_Target_Drug
}
\description{
A collection of actions related to drugs Targets
}
\keyword{datasets}
