% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Actions_Transporter_Drug}
\alias{Actions_Transporter_Drug}
\title{Drug Transporters related Actions}
\format{
a tibble with 53 observations and 2 variables:
\describe{
  \item{text}{describe related action}
  \item{parent_key}{Transporter key}
}
}
\source{
\url{https://docs.drugbankplus.com/xml/}
}
\usage{
Actions_Transporter_Drug
}
\description{
A collection of actions related to drugs Transporters
}
\keyword{datasets}
