% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AHFS_Codes_Drug}
\alias{AHFS_Codes_Drug}
\title{The American Hospital Formulary Service (AHFS) identifier for Drugs}
\format{
a tibble with 11 observations and 2 variables:
\describe{
  \item{text}{ahfs code}
  \item{parent_id}{drugbank id}
}
}
\source{
\url{https://docs.drugbankplus.com/xml/}
}
\usage{
AHFS_Codes_Drug
}
\description{
A list of the American Hospital Formulary Service (AHFS) identifier for drugs
}
\keyword{datasets}
