% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Articles_Target_Drug}
\alias{Articles_Target_Drug}
\title{Drug Target related Articles}
\format{
a tibble with 962 observations and 4 variables:
\describe{
  \item{ref-id}{Identifier for the article being referenced.
  This is unique across all reference types (books, links, article).}
  \item{pubmed-id}{The PubMed identifier for the article.}
  \item{citation}{Article citation in a standard format.}
  \item{parent_id}{target id}
}
}
\source{
\url{https://docs.drugbankplus.com/xml/}
}
\usage{
Articles_Target_Drug
}
\description{
A list of articles that were used as references for drugs target
}
\keyword{datasets}
