% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Articles_Transporter_Drug}
\alias{Articles_Transporter_Drug}
\title{Drug Transporter related Articles}
\format{
a tibble with 1283 observations and 4 variables:
\describe{
  \item{ref-id}{Identifier for the article being referenced.
  This is unique across all reference types (books, links, article).}
  \item{pubmed-id}{The PubMed identifier for the article.}
  \item{citation}{Article citation in a standard format.}
  \item{parent_id}{transporter id}
}
}
\source{
\url{https://docs.drugbankplus.com/xml/}
}
\usage{
Articles_Transporter_Drug
}
\description{
A list of articles that were used as references for drugs transporter
}
\keyword{datasets}
