% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Books_Drug}
\alias{Books_Drug}
\title{Drug related Books}
\format{
a tibble with 1 observation and 4 variables:
\describe{
  \item{ref-id}{Identifier for the article being referenced. 
  This is unique across all reference types (books, links, article).}
  \item{isbn}{ISBN identifying the textbook.}
  \item{citation}{A Textbook citation in a standard format.}
  \item{parent_key}{drug key}
}
}
\source{
\url{https://docs.drugbankplus.com/xml/}
}
\usage{
Books_Drug
}
\description{
A list of text books that were used as references for drugs
}
\keyword{datasets}
