% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Carriers_Drug}
\alias{Carriers_Drug}
\title{Targets / Enzymes / Carriers / Transporters}
\format{
a tibble with 13 observations and 6 variables:
\describe{
  \item{id}{Identifier for the record}
  \item{name}{related name}
  \item{organism}{Organism that the protein comes from.}
  \item{known_action}{Whether the pharmacological action of the drug is due
   to this taget interaction.}
  \item{position}{related position}
  \item{parent_key}{drug key}
}
}
\source{
\url{https://docs.drugbankplus.com/xml/}
}
\usage{
Carriers_Drug
}
\description{
Protein targets of drug action, enzymes that are inhibited/induced or 
involved in metabolism, and carrier or transporter proteins involved in 
movement of the drug across biological membranes.
}
\details{
Each of targets, enzymes, carriers and transporters contain one or more 
child elements tibbles
}
\keyword{datasets}
