% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19dbcand.R
\docType{package}
\name{covid19dbcand}
\alias{covid19dbcand}
\title{covid19dbcand: A data package that contains different datasets extracted 
from \strong{DrugBank} xml database.}
\description{
The datasets in `covid19dbcand` is related to 
[DrugBank](https://www.drugbank.ca/covid-19#drugs) selected drugs for
covid-19
}
\details{
The package is a smaller version from 
[dbdataset package](https://github.com/MohammedFCIS/dbdataset).

It is extracted using [dbparser](https://docs.ropensci.org/dbparser/) 

For more information kindly check the
reference/index 
(https://mohammedfcis.github.io/covid19dbcand/reference/index.html)
}
