% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Drugs_Pathway_Drug}
\alias{Drugs_Pathway_Drug}
\title{Pathway Drugs
Pathway Drugs
Each drug may have one or more pathway and vise versa}
\format{
a tibble with 33 observations and 3 variables:
\describe{
  \item{drugbank-id}{}
  \item{name}{drug name}
  \item{parent_id}{pathway id}
}
}
\source{
\url{https://docs.drugbankplus.com/xml/}
}
\usage{
Drugs_Pathway_Drug
}
\description{
Pathway Drugs
Pathway Drugs
Each drug may have one or more pathway and vise versa
}
\keyword{datasets}
