% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Enzymes_Pathway_Drug}
\alias{Enzymes_Pathway_Drug}
\title{Pathway Enzymes
Enzymes involved in this pathway.}
\format{
a tibble with 77 observations and 2 variables:
\describe{
  \item{text}{uniprot id}
  \item{parent_id}{pathway id}
}
}
\source{
\url{https://docs.drugbankplus.com/xml/}
}
\usage{
Enzymes_Pathway_Drug
}
\description{
Each drug may have one or more pathway
}
\keyword{datasets}
