% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Experimental_Properties_Drug}
\alias{Experimental_Properties_Drug}
\title{Drug Experimental Properties
Drug properties that have been experimentally proven.}
\format{
a tibble with 43 observations and 4 variables:
\describe{
  \item{kind}{Name of the property}
  \item{value}{Drug properties that have been experimentally proven}
  \item{source}{Reference to the source of this experimental data}
  \item{parent_key}{drugbank key}
}
}
\source{
\url{https://docs.drugbankplus.com/xml/}
}
\usage{
Experimental_Properties_Drug
}
\description{
Each drug may have one or more experimental property.
}
\keyword{datasets}
