% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Textbooks_Target_Drug}
\alias{Textbooks_Target_Drug}
\title{Drug Targets related Textbooks}
\format{
a tibble with 6 observations and 4 variables:
\describe{
  \item{ref-id}{Identifier for the article being referenced. 
  This is unique across all reference types (Textbooks, links, article).}
  \item{isbn}{ISBN identifying the textbook.}
  \item{citation}{A Textbook citation in a standard format.}
  \item{parent_key}{target key}
}
}
\source{
\url{https://docs.drugbankplus.com/xml/}
}
\usage{
Textbooks_Target_Drug
}
\description{
A list of text Textbooks that were used as references for drugs target
}
\keyword{datasets}
