% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Textbooks_Transporter_Drug}
\alias{Textbooks_Transporter_Drug}
\title{Drug Transporter related Textbooks}
\format{
a tibble with 4 observations and 4 variables:
\describe{
  \item{ref-id}{Identifier for the article being referenced.
  This is unique across all reference types (Textbooks, links, article).}
  \item{isbn}{ISBN identifying the textbook.}
  \item{citation}{A Textbook citation in a standard format.}
  \item{parent_id}{Transporter id}
}
}
\source{
\url{https://docs.drugbankplus.com/xml/}
}
\usage{
Textbooks_Transporter_Drug
}
\description{
A list of text Textbooks that were used as references for drugs transporters
}
\keyword{datasets}
